[catnip] Enhancement: Move to transport trait
