+++ src/rust/catnip/runtime/memory/mod.rs
@@ -16,7 +16,7 @@ pub use self::manager::MemoryManager;
 // Imports
 //==============================================================================
 
-use super::DPDKRuntime;
+use super::SharedDPDKRuntime;
 use crate::runtime::{
     fail::Fail,
     memory::{
@@ -31,7 +31,7 @@ use crate::runtime::{
 //==============================================================================
 
 /// Memory Runtime Trait Implementation for DPDK Runtime
-impl MemoryRuntime for DPDKRuntime {
+impl MemoryRuntime for SharedDPDKRuntime {
     /// Casts a [DPDKBuf] into an [demi_sgarray_t].
     fn into_sgarray(&self, buf: DemiBuffer) -> Result<demi_sgarray_t, Fail> {
         self.mm.into_sgarray(buf)
